lightning/pyproject.toml
[tool.poetry]
name = "lightning_baseline"
version = "0.1.0"
description = ""
authors = ["jemin7709 <jemin7709@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = {url="https://download.pytorch.org/whl/cu117/torch-2.0.1%2Bcu117-cp39-cp39-linux_x86_64.whl"}
torchvision = {url="https://download.pytorch.org/whl/cu117/torchvision-0.15.2%2Bcu117-cp39-cp39-linux_x86_64.whl"}
albumentations = "^1.3.1"
hydra-core = "^1.3.2"
lightning = "^2.0.3"
matplotlib = "^3.7.1"
ray = "^2.5.0"
segmentation-models-pytorch = "^0.3.3"
wandb = "^0.15.4"
pandas = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
